cmd/mev-boost: remove weird symlink
